group-members-page: Invalidate `canEditMembers` query after actions
In case the user edits their own membership, the `canEditMembers` query
needs to be invalidated so that we refresh the display based on the
user's current permissions.

If user removes themselves from the `owner` role, but still has other
roles, then the UI will change to the default list of members without
editing abilities.