Auto-prepend `/` to tile URLs
Previously, the code assumed that all URLs started with `/` and was
noted in featured-analyses.yaml. However, there was nothing to check
this. A `/` is perfectly valid for everything except for the source
check, which silently assumes everything that doesn't start with `/` is
maintained by Nextstrain. This is a bad assumption.

The source check could either (1) be updated to work with the error or
(2) throw a runtime error. I chose to go with (1) to make things easier
when updating featured analyses.