Remove PreviewOverlay when expanded
In all current usages of ExpandableTiles, it is possible for and
sometimes even expected for users to interact with tiles by clicking.
Having an invisible overlay on top of the last row of tiles with a
separate onClick behavior can be confusing for users expecting another
onClick behavior.

When the tiles are expanded, the only benefit to having PreviewOverlay
available is for the onClick behavior of hiding the tiles. This can
already be accomplished by the ArrowButton which also covers full width
and has a height of 1em.

This removes the opacity transition, which in my eyes does not make a
noticeable difference. The visually distinct part of the transition
action is on the height change to show more rows of tiles, not the
opacity here.