Fetch GroupMembersPage data with server action
The Page wrapper is a server side component that can use server side
actions to get Group data. The data are passed as props to the
GroupMembersPage, which is the client side component that renders the
styled components.

This removes the complexity of needing useState/useEffect to fetch the
Group data via API routes.