group-members-page: Redirect when user updates own roles
To force the user context from the server to be refreshed, the server
action has to redirect to page. Using `revalidatePath` is sufficient
for updating the page's data if the user is not changing their own
roles.

The redirect highlights a weird flicker in the UI on the load of the
page. First displays the "Loading" page as expected, then page flashes
without styles and finally shows the fully styled page.