Merge display name and tile lists
This update simplifies the YAML file by removing redundant
implementation details and should make content updates more
straightforward. The `coreGroupDisplayNames` and `coreTiles` lists had
overlapping information and are now combined into a single list of
pathogens.

Separate lists for display names and tiles are still used in the code,
derived from the merged list.

The key change is the tighter coupling between the display names shown
in the card titles and the names shown in the tiles.