Oropouche blog post: Clean up link outs
I like highlighting that there's a specific URL that corresponds to the viz shown in the screenshot. I've added these as "Available as..." to each figure legend.

Also, I've cleaned up some of these URLs. For example `p=grid` vs `p=full` will interact strongly with the user's current screen width, eg, you don't want a link to `p=grid` on a phone. Better to just leave these out.

Also, the M and S figure is showing separate trees rather than tanglegram.

I just removed `dmin=1986-01-01`. This is not what the figure is showing. Unfortunately we don't have a way to directly link to this zoomed clade. If we wanted this we could add an "outbreak" (or something else) clade label, but I don't think it's terribly important.

Also, I removed white-space in line endings.