Add stub 'Splash' component [#134]
Note: the '404' page uses the same `<Splash />` component as the home
page. I currently need to port the '404' page to the App Router, to
resolve some test failures, but I do not want to get bogged down into
porting the `<Splash />` component right now. Instead, I am adding
this stub, which should allow the 404 page to load and (most
importantly) satisfy the test suite.

The full `<Splash />` component will be fully ported later, at which
point the older version of the component will be removed.