Clarify why styles is not an allowed attribute
Add a reference and mention globals.css as an alternative. This is not
ideal since it adds unnecessary/unused styles for other pages, but this
is the easiest thing to do in absence of a more specific static CSS
file.

CSS modules cannot be easily used since the class names are dynamically
generated and meant for React/Next.js components.