Remove extension on ESM imports
This reverts commit "auspice-client: Append extension on ESM imports"
(7cde8399).

The reasoning in that commit was that extensions are necessary for
Webpack 5. However, this is not entirely accurate - they are only
required¹ when using the default for resolve.fullySpecified *and* using
"type": "module" package.json. The latter is not true in our case.

Notably, removing the exact path for
`auspice/src/components/markdownDisplay` allows the usage to be tolerant
of filename changes such as TypeScript conversion.

¹ <https://webpack.js.org/configuration/module/#resolvefullyspecified>