Initial move of Spinner component into App Router style [#1066]
Note: despite the placement of the component outside the
<ListResources> component, it is only used within that component.

I considered relocating it to
`static-site/components/list-resources/spinner.tsx` but elected to
maintain it outside of the <ListResources> component to better support
future re-use, which seems not unlikely.

Note also: the component is likely broken as of this commit, due to
this move; doing this as a distinct commit to better track the file
move.