Port /pathogens to App Router [#1066]
This moves the resource listing callback into a distinct file; this is
because Next.js Pages are limited in the types of methods that can be
exported from them.

The resource listing code was moved into a distinct method, which is
imported by the `<ListResources>` component, so that this page file
can remain a React Server Component, which allows for the Metadata API
to be used to generate content in the page `<head>` during SSR, prior
to hydration.