Port `/sars-cov-2/forecasts` to App Router [#1125]
* Move styles into CSS module
* Refactor code for consistency with other App Router pages
* Switch abstract section from <FocusParagraph> to
  <FocusParagraphNarrow> for consistency with other pages
* Add Metadata export

NOTE: This changes the width of the "abstract" section of the page so
that it matches other pages on the site; previously, the "abstract"
section of the `/sars-cov-2/forecasts` was wider than the 640px
default used on other pages.

Having everything use a consistent layout is an obvious maintenance
win; I picked the 640px variation because that's more common across
all the existing site pages.