Link to BA.2-derived viruses for 6m ncov builds
Circulating SARS-CoV-2 viruses are increasing evolved from the root of the tree. The default tree view should focus more on circulating viruses. A simple way to do this to have the default zoom be the most recent clade prior to the common ancestor of circulating viruses. At the moment this is clade 21L (lineage BA.2).

In the future we may be able to set these defaults as part of the Auspice JSON, but at the moment this isn't possible. So instead, this commit updates links to 6m builds to include the URL param that zooms to clade 21L.