endpoints/cli: Fix latent bug in pagination of releases
Extract just the URL from the Link header, not the whole {uri, rel}
object.

This bug is latent because the CLI has not yet exceeded 100 releases.
Lowering the per_page parameter to 10 exhibits the bug, which arises as
a "TypeError: Invalid URL" from passing an object instead of a string.