docs/api-restful: Add media types of resources for groups settings/memberships
Inferrable from examples and other doc, but good to spell it out here.
The current organization of this doc page, which is very focused on the
datasets/narratives resources, is starting to show its unsuitability for
our other resources.  It's ok for now, but we should be mindful as we
add more resources that stepping back a bit to reorganize the page could
be fruitful.