Use inclusion-based file target
Instead of exclusion-based, which is prone to accidentally linting
arbitrary files that one may have due to testing or an older build.

For static-site, this nicely aligns with the include config in the
TSConfig file (except for the Next.js-specific paths).

This approach can't be found in ESLint docs, which advertise a
'overrides.files' option that doesn't achieve exactly what's wanted
here. Instead, I was inspired by <https://stackoverflow.com/q/61720584>.