blog: Parse inline Markdown in post titles
This allows for basic styling of titles, like *emphasis*, **bold**, and
`code`.  The last is particularly useful for offsetting command names
from the rest of the text.

I reviewed all existing titles and sidebarTitles to confirm that no
rendering changes should occur and that this rendering change is safe to
apply retroactively.