Adjust global style and layout.txs to accomodate… [#1148]
…different index page layout.

The `padding: 0 25px` style that was applied to the `<main>` tag in
global.css is not needed (or wanted!) on the index page, only on
…every other page on the site. :facepalm:

Dealing with this is somewhat annoying, as React Server
Components (which is what all the porting work is moving towards)
aren't allowed to know what URL they're rendering — so we can't have a
parameterized check for "is this the index page".

I considered using multiple root layouts to support this, but that
made the filesystem layout fairly unintiuitive (and in testing it
ended up with some weird corner case bugs too).

Instead, I have opted to move the `padding` styling into a named
selector, and then wrap all the not-the-top-level-index-page pages in
a top-level `<div>` with that className.

Similarly, it was necessary to add a similar div around the <Footer>
component in layout.tsx, to bring things into alignment with the
former laytout.