[list-resources] separate fetching & display
of the ListResources component. These were previously somewhat separate
(you provided a callback to return data of a certain type) but we then
had a large amount of data munging within the component. It's much
cleaner & simpler to do all of this outside the component so that the
component can focus on the rendering of the processed data.