[list-resources] store filters as URL queries
which allows better linking to specific views. I didn't add a fallback
(so it'll be the default of `null`) but we add one if the loading is
noticeably slow.

The Suspense boundary is needed when using `useSearchParams()` - see
<https://nextjs.org/docs/messages/missing-suspense-with-csr-bailout>
and it must be inside the `<div ref={ref}>` element so that the
element width (`elWidth`) can be correctly measured.