[list-resources] fix modal scrolling
This removes the ability to scroll the background page (which was always
unintended), whilst allowing the contents of the modal to scroll as
needed. The current usage of the modal shouldn't overflow (although it
would on very small window sizes) but upcoming work to list all
intermediate data file versions will overflow.