scripts/get-resource-index-ref-matrix: remove extra conditional
The conditional to test for the presence of the `scripts/get-resource-index.js`
file was added for backwards compatibility with deployed apps that did not 
have the script yet. It has been over a year since then and all the deployed 
apps should have the script, so removing this extra conditional to simplify 
things.