fetch: Send a custom User-Agent for server-side requests
Doing so is neighborly and good practice: it helps external sites
identify traffic as ours and contact us if we're causing issues.  But it
also helps us more readily identify (and filter out or filter to) our
own traffic in S3 access logs (which is what prompted me to make this
change).

Use import.meta.resolve() and filesystem operations to get the
make-fetch-happen version at startup instead of an import
statement/function call with a {type: "json"} attribute because the
latter is still experimental in Node.js and thus issues a warning.
import.meta.resolve() requires Node.js 20.8, according to MDN¹, which
should have no impact since we expect a more recent 20.x version in
practice anyway.

¹ <https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/import.meta/resolve#browser_compatibility>