dev: Upgrade Jest from 27 → 30
Gains us import.meta.resolve() support in Jest, which is required now
that the previous commit added usage of that ESM feature.  Tests now
pass again instead of all throwing the error:

    TypeError: (intermediate value).resolve is not a function

I reviewed the upgrade guides for versions 28, 29, and 30¹ to check for
breaking changes and found none affecting us.  Points to Jest for that,
at least!

¹ <https://jest-archive-august-2023.netlify.app/docs/28.x/upgrading-to-jest28>
  <https://jestjs.io/docs/upgrading-to-jest29>
  <https://jestjs.io/docs/upgrading-to-jest30>