endpoints/charon: Fix dataset canonicalization to account for the current Source
Fixes a subtle bug where changing the dataset *in Auspice* to a
non-canonical path caused a redirection that switched sources, e.g.
staging → core.

For example, if you were on /staging/measles/genome and switched measles
to enterovirus, Auspice made a /charon/getDataset request for
/staging/enterovirus which *should* have canonicalized to
/staging/enterovirus/d68/genome but *was* canonicalized to
/enterovirus/d68/genome.  This was demonstrable by:

    $ curl -IL https://nextstrain.org/charon/getDataset?prefix=/staging/enterovirus -so /dev/null -w '%{url_effective}\n'
    https://nextstrain.org/charon/getDataset?prefix=enterovirus%2Fd68%2Fgenome

Note that the "prefix" query param was missing "staging/".