Expose Nextclade dataset reference trees under /nextclade/…
The new Nextclade source class and related classes in the
Source/Resource/Subresource framework provide access the "latest"
Nextclade dataset reference trees and resolve a myriad of supported
aliases.

The resource indexer is extended with a new "resource collection" for
Nextclade that essentially transforms the existing Nextclade indexes
into what our resource indexer expects.  This, in turns, fits into
existing resourceIndexer/ and src/resourceIndex.js code to provide
access to historical versions of Nextclade dataset reference trees.
Bumps the resource index version to v9 since changes were made to
resourceIndex/.

The static-site/app/nextclade/… files were largely copied from
static-site/app/staging/… and then modified to refer to the "nextclade"
source instead.  There is a lot of boilerplate and duplication.  But
that appears to be the way it's been done for other usages, and I don't
have time to make it better so close to the eve of my departure.  The
biggest differences are in the resources.tsx file.

Resolves: <https://github.com/nextstrain/nextstrain.org/issues/1156>