static-site[list-resources]: Fix crash in modal_draw() when bottoming out during radius calculations
Fixes the algorithm which calculates an ideal radius for the points on
the beeswarm plot by ensuring that beeswarmData, beeswarmHeight, and
radius are all set to their last computed values when bottoming out as
well as when topping out.  Simplifies state tracking instead of doubling
variables unnecessarily.

Previously when bottoming out on downward iterations, the algorithm
would leave beeswarmData unset, which caused errors like this from
inside d3:

    TypeError: can't access property Symbol.iterator, items is undefined

This was triggered when many data points were close together causing
beeswarm heights to soar.

Bottoming out will mean that some points in the plot are cut off, i.e.
above the top of the chart.  This seems preferable to a crash (or
unclickably tiny points).  Future improvements could dynamically scale
the chart height in such cases, or alternatively the chart width to
spread out points horizontally and make the longer timeline scrollable.

Resolves: <https://github.com/nextstrain/nextstrain.org/issues/1252>