[dependency] Remove "request"
This has been an on-again-off-again dependency within nextstrain.org.
The most recent addition was 62d17e1ba6f79e449cd3d75044003b8a8ef7f030,
although I can't see where that commit actually used the library.
There doesn't seem to be any usage in the current codebase.

The motivation for this was the deprecation of the entire library
<https://github.com/request/request/issues/3142> and the deprecation
of its sub-dependencies.