Set logo per resource group
Previously, a single logo reference was used for all resource listings –
see the removed TODO in resource-group.tsx.

Since this commit is a refactor without any functional changes, the
Nextstrain logo is still used for everything, even external resources on
the groups page. The following commit will change those to use
individual group logos.

The component is given in the data object to support asynchronous
fetches of external logos.