Define image element in list-resources
As mentioned in "Set logo per resource group" (bb73035), passing the
element as a component in the data object was meant to support
asynchronous fetches of external logos, but the next commit will not
take advantage of that.

This commit reverts back to defining the image element in the
list-resources component, but keeps it customizable by passing the image
source and alt text in the data object.