Use fa-users-solid as placeholder group logo
The main motivation is to stop displaying the Nextstrain logo next to
external group names, which sends a wrong message.

It's not been decided explicitly, but in a way we are using Font
Awesome's "Users" icon as the de facto logo for Nextstrain Groups. On
the splash page, it is displayed next to the "Nextstrain Groups" heading
and used under "Featured Analyses" as a placeholder logo for groups
without custom logos. This makes it natural to use the icon next to
external group names on the groups page.

It would be better to use group-specific logos (when available) and
create a more personalized logo for Nextstrain Groups to replace current
usage of the generic-looking fa-users-solid, but those improvements can
happen later.