Use custom group logos
For groups that have custom logos, using those logos in place of the
generic users icon gives a more personalized look to the page. The
generic icon is still used for groups without custom logos.

This is done with a new server-side function getGroupAvatars() to return
logos for all visible groups.