Move data fetching out of ListResources
ListResources was originally designed with a callback pattern where
parent components would specify a fetchResourceGroups function to be
called internally. This pattern prevented parent components from
controlling the data fetch, and led to duplicate fetches such as in
GroupListingPage.

This refactors the the flow to have parent components control the data
fetch and pass the result to ListResources. The other request in
GroupListingPage can now be re-used, and loading/error state management
is now controlled on the parent level meaning ListResources has a
simpler remit of listing resources.