Add choices for Heroku app input
The only supported values for this input are names of Heroku apps that
have a GitHub environment provisioned in the repo settings. Explicitly
defining the current valid choices¹ avoids the need to look up or
remember names, and prevents accidental typos.

Unfortunately, empty string / null is not allowed as a choice, so "none"
is used to represent no deployment.

¹ "dev: Allow manually-invoked deploys to a named Heroku app" (df780e7)