Set custom run name for deploys
This makes it easier to distinguish deploys from non-deploys on the
workflow run page.

Note: The condition representing the two possible paths to deployment
are duplicated between `run-name` and `jobs.vars`. Ideally, we could set
a top-level env var for the app name with a single condition, but
unfortunately it's not possible to access `env` in `run-name`.