Use dynamic routes for /groups pages
This reorganizes files from a flat structure with everything under a
catch-all [[...groups]] to a nested structure with one directory per URL
part.

A new catch-all page has been added for /groups/[group]/[...path] to
handle nonexistent resources.

A byproduct of dropping the custom top-level router is that the title
for /groups/* pages changes from "Scalable Sharing with Nextstrain
Groups - Nextstrain" to simply "Nextstrain". I'll address that in the
following commit.

Co-authored-by: John SJ Anderson <janders4@fredhutch.org>
Co-authored-by: Jover Lee <joverlee521@gmail.com>