Don't render Nextstrain logo for community
Setting the Nextstrain logo in the ListResources component and API made
sense back when it was only used for Nextstrain-maintained resources,
but that is no longer the case. The /groups page removed some hardcoding
by making the groupLogo customizable per group¹², but it left hardcoding
in the API, which is used by the new Nextclade datasets page.

This commit moves the hardcoding out of the API by adding a callback to
set the logo per resource group.

¹ "Set logo per resource group" (bb73035)
² "Define image element in list-resources" (3ef5ed9)