Mark CommunityRepoPage as Client Component
In practice, not marking it as such has worked fine because the only
usage is with a parent that is a Client Component, meaning it is
implicitly used as one as well.

Not marking it is fragile. Without "use client", using it in a parent
that is a Server Component (done in the next commit) results in a run
time error:

    You're importing a component that needs useEffect. It only works in a
    Client Component but none of its parents are marked with "use client",
    so they're Server Components by default.