Use dynamic routes for /community pages
This reorganizes files from a flat structure with everything under a
catch-all [[...community]] to a nested structure with one directory per
URL part.

Note that "narratives" is an optional part in the path. Such logic was
easily implemented in the custom router by using Array.shift(), but it
can't be encoded into dynamic routes. As a result, there is some
boilerplate duplication between community/[user]/ and
narratives/[user]/, but the bulk of the content remains in the top-level
community/*.tsx files.