Remove metadata/title prop from content pages
For each of the updated components in */content.tsx, the same value is
expected for the single prop. Some used metadata then extracted title,
while others used title directly. The inconsistency and overall
redundancy have been removed. This also removes the need for the custom
TitledMetadata type added in "Remove metadata.title type casts"
(13e2919).

Ideally, the title would be exported from an existing file, but that's
not possible with Next.js limitations. Storing in a separate constants
file does the job.