Move sort to comparison
This is motivated by an inconsistency where selecting multiple values
via the dropdown in non-alphabetical order stores values as sorted in
the URL search params, meaning the order will be changed upon page
refresh.

The change does not change that behavior (not sure why), but at least it
shows intent internally to indicate that sorting is only used for
comparison and not deliberately changing the order on users.