Restore group name filtering
The commit "Strip group name in resource list" (c7b44980) had the
unintended side effect of removing the group name from filter options.

This is a follow-up that keeps the visual change while allowing group
names to be used in filtering. It is accomplished by adding a
resource-level property to control which name parts are displayed, using
that to hide the first name part on the groups page.